Renamed Connections to ConnectionsTable
